Roll back mdlines -- it leads to stripping away trailing newlines
In effekt-standalone/Fig6 we could observe that omitting mdlines
leads to bugs. While previously, we generated

  ...%mdk\n
  \mdline{...}&

now we can observe

  ...%mdk&

which is clearly wrong, since the tab character & is now commented out.
The reason seems to be that somewhere the generated contents are stripped
(that is trailing whitespace, including newlines) is removed. If, after
generating a %mdk\n we also have a \mdline, then the \n is not trailing.

Not generating the %mdk at all also does not suffice, since then there
might be linebreaks. For instance in Theorem 4.4 same paper.

I leave in some debug information, which is used to figure out which
comment and mdline is generated where. Debugging can be picked up from
here at a later point.